Issue feat: update block list
